# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BackEnd to timer project",
        "title": "Go Task Chrono API",
        "termsOfService": "http://www.swagger.io/terms/",
        "contact": {
            "name": "Marcos Duarte",
            "url": "http://github.com/marcosduarte-dev/",
            "email": "pe.marcos30@gmail.com"
        },
        "license": {
            "name": "MarkDev License",
            "url": "http://github.com/marcosduarte-dev/"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/timers": {
            "get": {
                "description": "get all timers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "List timers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Timer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            },
            "post": {
                "description": "Create timers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "Create timer",
                "parameters": [
                    {
                        "description": "timer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TimerInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            }
        },
        "/timers/date/{date}": {
            "get": {
                "description": "Get timer by date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "Get timer by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Timer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            }
        },
        "/timers/task/{id}": {
            "get": {
                "description": "Get timer by task ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "Get timer by task ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Timer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            }
        },
        "/timers/{id}": {
            "get": {
                "description": "Get a timer by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "Get a timer",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "timer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Timer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a timer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "Update a timer",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "timer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "timer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TimerInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a timer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timers"
                ],
                "summary": "Delete a timer",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "timer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Return"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.TimerInputDTO": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "record_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "total_duration": {
                    "type": "integer"
                }
            }
        },
        "entity.Project": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.Return": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "entity.Task": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/entity.Project"
                },
                "project_id": {
                    "type": "string"
                }
            }
        },
        "entity.Timer": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "record_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "task": {
                    "$ref": "#/definitions/entity.Task"
                },
                "task_id": {
                    "type": "string"
                },
                "total_duration": {
                    "type": "integer"
                }
            }
        }
    }
}